Move tests from the test dir into wsgi_intercept/tests
This helps to insure the tests are included with the package
and to avoid conflicting with other packages.
