+++ tox.ini
@@ -6,7 +6,7 @@ envlist = py27,py33,py34,py35,pypy,pep8,docs
 [testenv]
 deps = .[testing]
 usedevelop = True
-commands = py.test test
+commands = py.test wsgi_intercept/tests
 passenv = WSGI_INTERCEPT_*
 
 [testenv:pep8]
